+++ src/main/kotlin/com/example/ggyunispring/domain/enum/DiaryType.kt
@@ -1,4 +1,4 @@
-package com.example.ggyunispring.common.enum
+package com.example.ggyunispring.domain.enum
 
 enum class DiaryType {
     YELLOW1, YELLOW2, BLUE1, BLUE2
