+++ src/main/kotlin/com/example/ggyunispring/domain/enum/Emotion.kt
@@ -1,4 +1,4 @@
-package com.example.ggyunispring.common.enum
+package com.example.ggyunispring.domain.enum
 
 enum class Emotion {
     MAD, SAD, PANIC, NORMAL, HAPPY
