+++ src/main/kotlin/com/example/ggyunispring/domain/exception/member/GoogleIdTokenException.kt
@@ -1,0 +1,4 @@
+package com.example.ggyunispring.domain.exception.member
+
+class GoogleIdTokenException : RuntimeException() {
+}
