+++ src/main/kotlin/com/example/ggyunispring/domain/exception/member/InvalidJwtTokenException.kt
@@ -1,0 +1,4 @@
+package com.example.ggyunispring.domain.exception.member
+
+class InvalidJwtTokenException: RuntimeException() {
+}
