--- src/main/kotlin/com/example/ggyunispring/exception/member/InvalidJwtTokenException.kt
@@ -1,4 +1,0 @@
-package com.example.ggyunispring.exception.member
-
-class InvalidJwtTokenException: RuntimeException() {
-}
